refactor: DataTable에서 WebSocket 분리
